+++ src/main/java/ai/labs/eddi/configs/dictionary/IRestExpression.java
@@ -15,7 +15,7 @@ import java.util.List;
  * @author ginccc
  */
 @Path("/expressions")
-@Tag(name = "Dictionary")
+@Tag(name = "Configuration / Dictionary", description = "Dictionary, expressions, and actions for NLP")
 public interface IRestExpression {
     @GET
     @Produces(MediaType.APPLICATION_JSON)